add tag and tag-list components, rename the old tag-list component to tag-filter
