+++ frontend/.storybook/addons.ts
@@ -1,3 +1,4 @@
-import '@storybook/addon-viewport/register'
+import '@storybook/addon-viewport/register';
 import '@storybook/addon-knobs/register';
 import '@storybook/addon-console';
+import '@storybook/addon-actions/register';
